requirements.txt
rendercv[full]>=2.3
pyyaml>=6.0
python-dateutil>=2.8
